backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2
google-auth==2.26.2
google-auth-oauthlib==1.2.0
psycopg2-binary
python-dotenv==1.0.0
yfinance==0.2.35
pytz==2024.1
